Turn the solution and main mod into API and create Example mod that implements the API.
