+++ BossRush/BossRushPlayer.cs
@@ -1,6 +1,7 @@
 ﻿using Terraria;
 using Terraria.DataStructures;
 using Terraria.ModLoader;
+using BRS = BossRush.BossRushSystem;
 
 namespace BossRush;
 
@@ -8,7 +9,7 @@ public class BossRushPlayer : ModPlayer
 {
     public override void UpdateDead()
     {
-        if (BossRushSystem.IsBossRushActive())
+        if (BRS.I.IsBossRushActive)
         {
             Player.respawnTimer = 10.ToFrames();
         }
@@ -18,26 +19,26 @@ public class BossRushPlayer : ModPlayer
 
     public override void PostUpdate()
     {
-        if (BossRushSystem.IsBossRushActive())
+        if (BRS.I.IsBossRushActive)
         {
-            BossRushSystem.I.CurrentBossData?.PlaceContext?.ForceBiomeFunction?.Invoke(Player);
+            BRS.I.CurrentBossData?.PlaceContext?.ForceBiomeFunction?.Invoke(Player);
         }
     }
 
     public override void Kill(double damage, int hitDirection, bool pvp, PlayerDeathReason damageSource)
     {
-        if (BossRushSystem.IsBossRushActive())
+        if (BRS.I.IsBossRushActive)
         {
-            BossRushSystem.I.TrackPlayerDeaths();
+            BRS.I.TrackPlayerDeaths();
         }
     }
 
     public override void PlayerDisconnect()
     {
-        if (BossRushSystem.IsBossRushActive())
+        if (BRS.I.IsBossRushActive)
         {
             Player.active = false;
-            BossRushSystem.I.TrackPlayerDeaths();
+            BRS.I.TrackPlayerDeaths();
         }
     }
 }
