+++ BossRush/NPCs/RandomMobs.cs
@@ -1,5 +1,6 @@
 ﻿using Terraria;
 using Terraria.ModLoader;
+using BRS = BossRush.BossRushSystem;
 
 namespace BossRush.NPCs;
 
@@ -9,7 +10,7 @@ public class RandomMobs : GlobalNPC
 
     public override void EditSpawnRate(Player player, ref int spawnRate, ref int maxSpawns)
     {
-        if (BossRushSystem.IsBossRushActive())
+        if (BRS.I.IsBossRushActive)
         {
             // Using this might have side-effects.
             // maxSpawns = 0;
