+++ ExampleBossRush/ExampleBossRush.cs
@@ -1,0 +1,7 @@
+using Terraria.ModLoader;
+
+namespace ExampleBossRush;
+
+public class ExampleBossRush : Mod
+{
+}
