+++ ExampleBossRush/ExampleBossRush.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<!-- Import tModLoader mod properties -->
-	<Import Project="..\tModLoader.targets" />
+	<Import Project="..\..\tModLoader.targets" />
 
 	<!-- General -->
 	<PropertyGroup>
@@ -9,5 +9,8 @@
 	</PropertyGroup>
 
 	<!-- References -->
+	<ItemGroup>
+	  <ProjectReference Include="..\BossRush\BossRush.csproj" />
+	</ItemGroup>
 
 </Project>
