+++ ExampleBossRush/Localization/en-US_Mods.ExampleBossRush.hjson
@@ -1,0 +1,2 @@
+# This file will automatically update with entries for new content after a build and reload.
+
