+++ ExampleBossRush/description.txt
@@ -1,0 +1,1 @@
+Modify this file with a description of your mod.
