--- build.txt
@@ -1,3 +1,0 @@
-displayName = Boss Rush
-author = tieeeeen1994
-version = 0.1
